+++ SNESMimicTypeDealy/Components/Animation/AnimatorComponent.h
@@ -27,7 +27,7 @@ namespace GAME_NAME
 				/// Update the speed multiplier. (Lower values result in longer frames.)
 				/// </summary>
 				/// <param name="speedMult">How many times faster to make the animation.</param>
-				inline void SetSpeedMult(float speedMult)
+				inline void SetSpeedMult(double speedMult)
 				{
 					m_speedMult = speedMult;
 				}
@@ -45,9 +45,9 @@ namespace GAME_NAME
 				std::vector<std::shared_ptr<Animation>> m_animations;					//List of all possible animations this object can have.
 				int8_t m_currentAnimation = -1;											//Current animation to use.
 
-				float m_speedMult = 1.f;												//Speed multiplier for controling animation speed based on events. (Lower values result in longer frames.)
+				double m_speedMult = 1.0;												//Speed multiplier for controling animation speed based on events. (Lower values result in longer frames.)
 
-				float m_tick = 0.f;														//Storage for when to advance to the next frame.
+				double m_tick = 0.0;													//Storage for when to advance to the next frame.
 			};
